# apps/python/openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OAP LangGraph Runtime",
    "version": "0.1.0",
    "description": "\nRobyn-based LangGraph-compatible runtime for Open Agent Platform.\n\nThis API provides endpoints for managing assistants, threads, runs, and the store.\nIt implements the LangGraph API specification for compatibility with LangGraph clients.\n"
  },
...
  "paths": {
...
    "/mcp/": {
      "post": {
        "tags": [
          "MCP"
        ],
        "summary": "MCP Post",
        "description": "Implemented according to the Streamable HTTP Transport specification.\nSends a JSON-RPC 2.0 message to the server.\n\n- **Request**: Provide an object with `jsonrpc`, `id`, `method`, and optional `params`.\n- **Response**: Returns a JSON-RPC response or acknowledgment.\n\n**Supported Methods:**\n- `initialize` - Client handshake with capabilities\n- `tools/list` - List available tools (returns the LangGraph agent)\n- `tools/call` - Execute the agent with a message\n\n**Notes:**\n- Stateless: Sessions are not persisted across requests.\n",
        "operationId": "post_mcp",
        "parameters": [
          {
            "name": "Accept",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "application/json",
                "application/json, text/event-stream"
              ]
            },
            "description": "Accept header should include 'application/json'."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "jsonrpc": {
                    "type": "string",
                    "enum": [
                      "2.0"
                    ],
                    "description": "JSON-RPC version (must be '2.0')"
                  },
                  "id": {
                    "oneOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "integer"
                      }
                    ],
                    "description": "Request ID for correlating responses"
                  },
                  "method": {
                    "type": "string",
                    "description": "Method to invoke (e.g., 'initialize', 'tools/list', 'tools/call')"
                  },
                  "params": {
                    "type": "object",
                    "description": "Method parameters"
                  }
                },
                "required": [
                  "jsonrpc",
                  "method"
                ]
              },
              "example": {
                "jsonrpc": "2.0",
                "id": "1",
                "method": "initialize",
                "params": {
                  "clientInfo": {
                    "name": "test_client",
                    "version": "1.0.0"
                  },
                  "protocolVersion": "2024-11-05",
                  "capabilities": {}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful JSON-RPC response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jsonrpc": {
                      "type": "string",
                      "enum": [
                        "2.0"
                      ]
                    },
                    "id": {
                      "oneOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "integer"
                        }
                      ]
                    },
                    "result": {
                      "type": "object"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "integer"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "202": {
            "description": "Notification accepted; no content body"
          },
          "400": {
            "description": "Bad request: invalid JSON or message format"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "tags": [
          "MCP"
        ],
        "summary": "MCP Get",
        "description": "Implemented according to the Streamable HTTP Transport specification. GET is not supported (streaming not implemented).",
        "operationId": "get_mcp",
        "responses": {
          "405": {
            "description": "GET method not allowed; streaming not supported"
          }
        }
      },
      "delete": {
        "tags": [
          "MCP"
        ],
        "summary": "Terminate Session",
        "description": "Implemented according to the Streamable HTTP Transport specification. Since the server is stateless, there are no sessions to terminate.",
        "operationId": "delete_mcp",
        "responses": {
          "404": {
            "description": "Session not found (server is stateless)"
          }
        }
      }
    },
    "/runs/crons": {
      "post": {
        "tags": [
          "Crons"
        ],
        "summary": "Create Cron",
        "description": "Create a cron to schedule runs on new threads.",
        "operationId": "create_cron_runs_crons_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CronCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cron"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/runs/crons/search": {
      "post": {
        "tags": [
          "Crons"
        ],
        "summary": "Search Crons",
        "description": "Search all active crons.",
        "operationId": "search_crons_runs_crons_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CronSearch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Cron"
                  },
                  "type": "array",
                  "title": "Response Search Crons"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/runs/crons/count": {
      "post": {
        "tags": [
          "Crons"
        ],
        "summary": "Count Crons",
        "description": "Get the count of crons matching the specified criteria.",
        "operationId": "count_crons_runs_crons_count_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CronCountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "title": "Count"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/runs/crons/{cron_id}": {
      "delete": {
        "tags": [
          "Crons"
        ],
        "summary": "Delete Cron",
        "description": "Delete a cron by ID.",
        "operationId": "delete_cron_runs_crons__cron_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cron Id"
            },
            "name": "cron_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Error message describing what went wrong."
          }
        },
        "title": "ErrorResponse",
        "description": "Standard error response format."
      },
      "Config": {
        "type": "object",
        "properties": {
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "description": "Tags for categorizing the run."
          },
          "recursion_limit": {
            "type": "integer",
            "title": "Recursion Limit",
            "description": "Maximum recursion depth for the graph.",
            "default": 25
          },
          "configurable": {
            "type": "object",
            "title": "Configurable",
            "description": "Configurable parameters for the graph."
          }
        },
        "title": "Config"
      },
...
      "Cron": {
        "type": "object",
        "required": [
          "cron_id",
          "thread_id",
          "schedule",
          "created_at",
          "updated_at",
          "payload"
        ],
        "properties": {
          "cron_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cron Id",
            "description": "The ID of the cron."
          },
          "assistant_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "title": "Assistant Id",
            "description": "The ID of the assistant."
          },
          "thread_id": {
            "type": "string",
            "format": "uuid",
            "title": "Thread Id",
            "description": "The ID of the thread."
          },
          "end_time": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "title": "End Time",
            "description": "The end date to stop running the cron."
          },
          "schedule": {
            "type": "string",
            "title": "Schedule",
            "description": "The schedule to run, cron format."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "The time the cron was created."
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "The last time the cron was updated."
          },
          "user_id": {
            "type": [
              "string",
              "null"
            ],
            "title": "User Id",
            "description": "The ID of the user."
          },
          "payload": {
            "type": "object",
            "title": "Payload",
            "description": "The run payload to use for creating new run."
          },
          "next_run_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "title": "Next Run Date",
            "description": "The next run date of the cron."
          },
          "metadata": {
            "type": "object",
            "title": "Metadata",
            "description": "The cron metadata."
          }
        },
        "title": "Cron",
        "description": "Represents a scheduled task."
      },
      "CronCreate": {
        "type": "object",
        "required": [
          "assistant_id",
          "schedule"
        ],
        "properties": {
          "schedule": {
            "type": "string",
            "title": "Schedule",
            "description": "The cron schedule to execute this job on."
          },
          "end_time": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "title": "End Time",
            "description": "The end date to stop running the cron."
          },
          "assistant_id": {
            "type": "string",
            "title": "Assistant Id",
            "description": "The assistant ID or graph name to run."
          },
          "input": {
            "anyOf": [
              {
                "items": {
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "object"
              }
            ],
            "title": "Input",
            "description": "The input to the graph."
          },
          "metadata": {
            "type": "object",
            "title": "Metadata",
            "description": "Metadata to assign to the cron job runs."
          },
          "config": {
            "$ref": "#/components/schemas/Config"
          },
          "context": {
            "type": "object",
            "title": "Context",
            "description": "Static context added to the assistant."
          },
          "webhook": {
            "type": "string",
            "format": "uri-reference",
            "title": "Webhook",
            "description": "Webhook to call after LangGraph API call is done."
          },
          "interrupt_before": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "*"
                ]
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ],
            "title": "Interrupt Before",
            "description": "Nodes to interrupt immediately before they get executed."
          },
          "interrupt_after": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "*"
                ]
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ],
            "title": "Interrupt After",
            "description": "Nodes to interrupt immediately after they get executed."
          },
          "on_run_completed": {
            "type": "string",
            "enum": [
              "delete",
              "keep"
            ],
            "default": "delete",
            "title": "On Run Completed",
            "description": "What to do with the thread after the run completes."
          }
        },
        "title": "CronCreate",
        "description": "Payload for creating a stateless cron job."
      },
      "CronSearch": {
        "type": "object",
        "properties": {
          "assistant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assistant Id",
            "description": "The assistant ID to filter by."
          },
          "thread_id": {
            "type": "string",
            "format": "uuid",
            "title": "Thread Id",
            "description": "The thread ID to search for."
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 10,
            "minimum": 1,
            "maximum": 1000,
            "description": "The maximum number of results to return."
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "default": 0,
            "minimum": 0,
            "description": "The number of results to skip."
          },
          "sort_by": {
            "type": "string",
            "title": "Sort By",
            "default": "created_at",
            "enum": [
              "cron_id",
              "assistant_id",
              "thread_id",
              "next_run_date",
              "end_time",
              "created_at",
              "updated_at"
            ],
            "description": "The field to sort by."
          },
          "sort_order": {
            "type": "string",
            "title": "Sort Order",
            "default": "desc",
            "enum": [
              "asc",
              "desc"
            ],
            "description": "The order to sort by."
          },
          "select": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "cron_id",
                "assistant_id",
                "thread_id",
                "on_run_completed",
                "end_time",
                "schedule",
                "created_at",
                "updated_at",
                "user_id",
                "payload",
                "next_run_date",
                "metadata"
              ]
            },
            "title": "Select",
            "description": "Specify which fields to return."
          }
        },
        "title": "CronSearch",
        "description": "Payload for listing crons."
      },
      "CronCountRequest": {
        "type": "object",
        "properties": {
          "assistant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assistant Id",
            "description": "The assistant ID to search for."
          },
          "thread_id": {
            "type": "string",
            "format": "uuid",
            "title": "Thread Id",
            "description": "The thread ID to search for."
          }
        },
        "title": "CronCountRequest",
        "description": "Payload for counting crons."
      },
...
    }
  }
}